+++ .eslintrc.json
@@ -69,7 +69,7 @@
         "devDependencies": [
           "**/test/**",
           "**/build-tools/**",
-          "src/**/__test__/**/*",
+          "src/**/__tests__/**/*",
           "src/**/__mocks__/**/*",
           "src/**/*.test.ts",
           "src/**/*.spec.ts",
@@ -192,12 +192,14 @@
     },
     {
       "files": [
-        "**/__test__/**/*.[jt]s",
-        "**/__test__/**/*.[jt]sx",
-        "*.spec.[jt]s",
+        "**/__tests__/**/*.[jt]s",
+        "**/__mocks__/**/*.[jt]s",
         "*.test.[jt]s",
-        "*.spec.[jt]sx",
-        "*.test.[jt]sx"
+        "*.spec.[jt]s",
+        "**/__tests__/**/*.[jt]sx",
+        "**/__mocks__/**/*.[jt]sx",
+        "*.test.[jt]sx",
+        "*.spec.[jt]sx"
       ],
       "plugins": [
         "jest-dom",
