+++ .projenrc.ts
@@ -4,19 +4,9 @@ import { Nvmrc } from "projen-nvm";
 
 const PACKAGE_NAME = "use-set-as-state";
 
-const DEV_AND_PEER_DEPENDENCIES = [
-    "immer",
-    "react",
-    "react-dom",
-    "use-immer-produce",
-];
+const DEV_AND_PEER_DEPENDENCIES = ["immer", "use-immer-produce"];
 const DEV_DEPENDENCIES = [
     ...DEV_AND_PEER_DEPENDENCIES,
-    "@testing-library/jest-dom",
-    "@testing-library/react",
-    "@types/react",
-    "@types/react-dom",
-    "jest-environment-jsdom",
     "dkershner6-projen",
     "projen-nvm",
 ];
@@ -37,11 +27,8 @@ const project = new Node20ReactTypeScriptProject({
     repository: `git+https://github.com/dkershner6/${PACKAGE_NAME}.git`,
     projenrcTs: true,
 
-    // deps: [],                /* Runtime dependencies of this module. */
-    // description: undefined,  /* The description is just a string that helps people understand the purpose of the package. */
     devDeps: DEV_DEPENDENCIES,
     peerDeps: PEER_DEPENDENCIES,
-    // packageName: undefined,  /* The "name" in package.json. */
 
     release: true,
     releaseToNpm: true,
@@ -49,8 +36,6 @@ const project = new Node20ReactTypeScriptProject({
 
     docgen: true,
 });
-
-project.tsconfig?.addExclude("**/__test__/**/*");
 
 new Nvmrc(project);
