+++ package.json
@@ -44,7 +44,7 @@
     "@typescript-eslint/eslint-plugin": "^6",
     "@typescript-eslint/parser": "^6",
     "constructs": "^10.0.0",
-    "dkershner6-projen": "^0.0.21",
+    "dkershner6-projen": "^0.0.23",
     "eslint": "^8",
     "eslint-config-prettier": "^9.1.0",
     "eslint-import-resolver-typescript": "^3.6.1",
@@ -64,7 +64,7 @@
     "jest-junit": "^15",
     "prettier": "^3.1.1",
     "projen": "^0.78.5",
-    "projen-nvm": "^0.0.30",
+    "projen-nvm": "^0.0.32",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "standard-version": "^9",
